+++ gamma_dashboard/static/dashboard/js/app/src/dashboard/constants.js
@@ -4,5 +4,7 @@ export const WIDGETS = {
 };
 
 export const SHOW_STATUS_BLOCK = true;
+export const SHOW_AVATAR_BLOCK = false;
+export const SHOW_VAULT_BLOCK = false;
 
 export const MAIN_CONTENT_ID = 'dashboard-page-title';